+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Support downloading Swedish data (#96).
 * How attribute is now present in Czech data (#102).
+* Use `withr` to prevent files being left in temporary directories (#98)
 
 # getRad 0.2.0